Bio.Phylogeny.PhyBin

Description

This module contains the code that does the tree normalization and binning. It's the heart of the prgoram.

Documentation

driver :: PhyBinConfig -> IO ()

Driver to put all the pieces together (parse, normalize, bin)

binthem :: [FullTree DefDecor] -> BinResults StandardDecor

The binning function. Takes labeled trees, classifies labels into equivalence classes.

normalize :: AnnotatedTree -> AnnotatedTree

This is it, here's the routine that transforms a tree into normal form. This relies HEAVILY on lazy evaluation.

annotateWLabLists :: NewickTree DefDecor -> AnnotatedTree

Add the metadata that is used for binning

unitTests :: Test

acquireTreeFiles :: [String] -> IO [String]

Expand out directories to find all the tree files.

deAnnotate :: FullTree StandardDecor -> FullTree DefDecor

Take the extra annotations away. Inverse of annotateWLabLists.

retrieveHighlights :: (String -> String) -> LabelTable -> [FilePath] -> IO [[NewickTree ()]]

Parse trees in addition to the main inputs (for --highlight).

matchAnyHighlight :: [[NewickTree ()]] -> NewickTree () -> Bool

Create a predicate that tests trees for consistency with the set of --highlight (consensus) trees.