Bio.Phylogeny.PhyBin.Binning

Description

This module contains the code that does the tree normalization and binning.

Binning and normalization

binthem :: [FullTree DefDecor] -> BinResults StandardDecor

The binning function. Takes labeled trees, classifies labels into equivalence classes.

normalize :: AnnotatedTree -> AnnotatedTree

This is it, here's the routine that transforms a tree into normal form. This relies HEAVILY on lazy evaluation.

normalizeFT :: FullTree StandardDecor -> FullTree StandardDecor

A version lifted to operate over full trees.

annotateWLabLists :: NewickTree DefDecor -> AnnotatedTree

Add the metadata that is used for binning

deAnnotate :: FullTree StandardDecor -> FullTree DefDecor

Take the extra annotations away. Inverse of annotateWLabLists.

newtype OneCluster a

When binning, the members of a OneCluster are isomorphic trees. When clustering based on robinson-foulds distance they are merely similar trees.

Constructors

OneCluster

Fields

clustMembers :: [FullTree a]

type BinResults a = Map StrippedTree (OneCluster a)

Index the results of binning by topology-only stripped trees that have their decorations removed.

type StrippedTree = NewickTree Int

Ignore metadata (but keep weights) for the purpose of binning

Utilities and unit tests

get_weight :: AnnotatedTree -> Int

unitTests :: Test

anonymize_annotated :: AnnotatedTree -> StrippedTree

For binning. Remove branch lengths and labels but leave weights.