Bio.Phylogeny.PhyBin.CoreTypes

Tree and tree decoration types

data NewickTree a

Even though the Newick format allows it, here we ignore interior node labels. (They are not commonly used.)

Note that these trees are rooted. The normalize function ensures that a single, canonical rooted representation is chosen.

Constructors

NTLeaf a !Label

NTInterior a [NewickTree a]

type DefDecor = (Maybe Int, BranchLen)

The barebones default decorator for NewickTrees contains BOOTSTRAP and BRANCHLENGTH. The bootstrap values, if present, will range in [0..100]

data StandardDecor

The standard decoration includes everything in DefDecor plus some extra cached data:

branch length from parent to this node (2) bootstrap values for the node

subtree weights for future use (defined as number of LEAVES, not counting intermediate nodes) (4) sorted lists of labels for symmetry breaking

Constructors

StandardDecor

Fields

branchLen :: BranchLen

bootStrap :: Maybe Int

subtreeWeight :: Int

sortedLabels :: [Label]

type AnnotatedTree = NewickTree StandardDecor

Additionally includes some scratch data that is used by the binning algorithm.

data FullTree a

A common type of tree contains the standard decorator and also a table for restoring the human-readable node names.

Constructors

FullTree

Fields

treename :: TreeName

labelTable :: LabelTable

nwtree :: NewickTree a

data ClustMode

Constructors

BinThem

ClusterThem

Fields

linkage :: Linkage

type TreeName = String

Tree operations

displayDefaultTree :: FullTree DefDecor -> Doc

Display a tree WITH the bootstrap and branch lengths.

displayStrippedTree :: FullTree a -> Doc

treeSize :: NewickTree a -> Int

How many nodes (leaves and interior) are contained in a NewickTree?

numLeaves :: NewickTree a -> Int

This counts only leaf nodes, which should include all taxa.

liftFT :: (NewickTree t -> NewickTree a) -> FullTree t -> FullTree a

get_dec :: NewickTree t -> t

set_dec :: b -> NewickTree a -> NewickTree b

get_children :: NewickTree t -> [NewickTree t]

map_labels :: (Label -> Label) -> NewickTree a -> NewickTree a

Apply a function to all the *labels* (leaf names) in a tree.

all_labels :: NewickTree t -> [Label]

Return all the labels contained in the tree.

foldIsomorphicTrees :: ([a] -> b) -> [NewickTree a] -> NewickTree b

This function allows one to collapse multiple trees while looking only at the horizontal slice of all the annotations *at a given position* in the tree.

Isomorphic must apply both to the shape and the name labels or it is an error to apply this function.

Utilities specific to StandardDecor:

avg_branchlen :: HasBranchLen a => [NewickTree a] -> Double

Average branch length across all branches in all all trees.

get_bootstraps :: NewickTree StandardDecor -> [Int]

Retrieve all the bootstraps values actually present in a tree.

Command line config options

data PhyBinConfig

Due to the number of configuration options for the driver, we pack them into a record.

Constructors

PBC

Fields

verbose :: Bool

num_taxa :: Int

name_hack :: String -> String

output_dir :: String

inputs :: [String]

do_graph :: Bool

do_draw :: Bool

clust_mode :: ClustMode

highlights :: [FilePath]

show_trees_in_dendro :: Bool

show_interior_consensus :: Bool

use_hashrf :: Bool

print_rfmatrix :: Bool

dist_thresh :: Maybe Int

branch_collapse_thresh :: Maybe Double

Branches less than this length are collapsed.

default_phybin_config :: PhyBinConfig

The default phybin configuration.

General helpers

type Label = Int

type LabelTable = Map Label String

Map labels back onto meaningful names.

Experimenting with abstracting decoration operations

class HasBranchLen a where

Methods

getBranchLen :: a -> BranchLen