Bio.Phylogeny.PhyBin.Parser

Documentation

newick_parser :: NameHack -> Parser TempTree

This parser ASSUMES that whitespace has been prefiltered from the input.

parseNewick :: LabelTable -> NameHack -> String -> ByteString -> (LabelTable, [NewickTree DefDecor])

Parse a bytestring into a NewickTree with branch lengths. The first argument is file from which the data came and is just for better error messages.

parseNewicks :: NameHack -> [(String, ByteString)] -> (LabelTable, [FullTree DefDecor])

A version which takes in-memory trees as ByteStrings.

parseNewickFiles :: NameHack -> [String] -> IO (LabelTable, [FullTree DefDecor])

Parse a list of trees, starting with an empty map of labels and accumulating a final map.

unitTests :: Test