Bio.Phylogeny.PhyBin.RFDistance

Types

type DenseLabelSet = IntSet

Dense sets of taxa, aka Bipartitions or BiPs We assume that taxa labels have been mapped onto a dense, contiguous range of integers [0,N).

Rule: Bipartitions are really two disjoint sets. But as long as the parent set (the union of the partitions, aka all taxa) then a bipartition can be represented just by *one* subset. Yet we must choose WHICH subset for consistency. We use the rule that we always choose the SMALLER. Thus the DenseLabelSet should always be half the size or less, compared to the total number of taxa.

A set that is more than a majority of the taxa can be normalized by flipping, i.e. taking the taxa that are NOT in that set.

type DistanceMatrix = Vector (Vector Int)

Bipartition (Bip) utilities

allBips :: NewickTree a -> Set DenseLabelSet

Get all non-singleton BiPs implied by a tree.

foldBips :: Monoid m => (DenseLabelSet -> m) -> NewickTree a -> m

dispBip :: LabelTable -> DenseLabelSet -> String

Print a BiPartition in a pretty form

consensusTree :: Int -> [NewickTree a] -> NewickTree ()

Take only the bipartitions that are agreed on by all trees.

bipsToTree :: Int -> Set DenseLabelSet -> NewickTree ()

Convert from bipartitions BACK to a single tree.

filterCompatible :: NewickTree a -> [NewickTree b] -> [NewickTree b]

Which of a set of trees are compatible with a consensus?

compatibleWith :: NewickTree a -> NewickTree b -> Bool

Is a tree compatible with a consensus? This is more efficient if partially applied then used repeatedly.

ADT for dense sets

mkSingleDense :: Int -> Label -> DenseLabelSet

mkEmptyDense :: Int -> DenseLabelSet

bipSize :: DenseLabelSet -> Int

denseUnions :: Int -> [DenseLabelSet] -> DenseLabelSet

denseDiff :: IntSet -> IntSet -> IntSet

invertDense :: Int -> DenseLabelSet -> DenseLabelSet

Assume that total taxa are 0..N-1, invert membership:

markLabel :: Label -> DenseLabelSet -> DenseLabelSet

Methods for computing distance matrices

distanceMatrix :: [NewickTree a] -> (DistanceMatrix, Vector (Set DenseLabelSet))

Returns a triangular distance matrix encoded as a vector. Also return the set-of-BIPs representation for each tree.

hashRF :: forall dec. Int -> [NewickTree dec] -> IO DistanceMatrix2

This version slices the problem a different way. A single pass over the trees populates the table of bipartitions. Then the table can be processed (locally) to produce (non-localized) increments to a distance matrix.

Output

printDistMat :: Handle -> Vector (Vector Int) -> IO ()